Moving redirect expectation functions into its own helper, and replacing the spec functions that were fulfilling the same purpose.
With a few other tweaks that came up in code review.
